requirements.txt
pandas
scikit-learn
flask
streamlit
joblib
requests
paho-mqtt
kafka-python